pyproject.toml
[tool.poetry]
name = "ps-app"
version = "0.1.0"
description = ""
authors = ["Jesse Egbosionu <dev@jesseinit.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
elasticsearch-dsl = "^8.15.3"
django-elasticsearch-dsl = "^8.0"
djangorestframework = "^3.15.2"
elasticsearch = "^8.15.1"
django-elasticsearch-dsl-drf = "^0.22.5"
factory-boy = "^3.3.1"
pytz = "^2024.2"
django-cte = "^1.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
